+++ sample/src/androidTest/java/com/zhuinden/espressohelpersample/ExampleInstrumentedTest.kt
@@ -2,10 +2,9 @@
 package com.zhuinden.espressohelpersample
 
 import android.support.constraint.ConstraintLayout
-import android.support.test.rule.ActivityTestRule
+import android.support.test.espresso.intent.rule.IntentsTestRule
 import android.support.test.runner.AndroidJUnit4
 import com.zhuinden.espressohelper.*
-import com.zhuinden.espressohelper.conditionwatcher.ConditionWatcher
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -21,17 +20,16 @@ class ExampleInstrumentedTest {
 
     @JvmField
     @field:Rule
-    var rule = ActivityTestRule(MainActivity::class.java)
+    var rule = IntentsTestRule(MainActivity::class.java)
 
     @Test
     fun clickButton() {
         R.id.container.checkIsAssignableFrom<ConstraintLayout>()
-        R.id.hello_world.checkHasText("Hello World!")
+        R.id.helloWorld.checkHasText(R.string.hello_world)
         R.id.username.performTypeText("This is a test!")
-        R.id.username.performCloseSoftKeyboard()
         R.id.password.performTypeText("hunter2")
-        R.id.password.performCloseSoftKeyboard()
         R.id.button.performClick()
-        ConditionWatcher.waitForCondition { true }
+        checkNextActivity<SecondActivity>()
+        R.id.secondText.checkHasText("Well done!")
     }
 }