+++ sample/src/main/java/com/zhuinden/espressohelpersample/MainActivity.kt
@@ -2,11 +2,18 @@ package com.zhuinden.espressohelpersample
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
+import kotlinx.android.synthetic.main.activity_main.*
+import org.jetbrains.anko.sdk15.listeners.onClick
+import org.jetbrains.anko.startActivity
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+        button.onClick {
+            startActivity<SecondActivity>()
+        }
     }
 }