+++ sample/src/main/java/com/zhuinden/espressohelpersample/SecondActivity.kt
@@ -1,0 +1,11 @@
+package com.zhuinden.espressohelpersample
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+
+class SecondActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_second)
+    }
+}