+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Espresso-Helper 0.1.1 (2018-05-XX)
+
+- Add `performActionOnRecyclerHolderItem`, `performActionOnRecyclerItem`, `performActionOnRecyclerItemAtPosition`
+
+- Add `checkNextActivity<T>`.
+
+- Add `closeSoftKeyboard()` to `typeText`, `clearText`, `typeTextIntoFocusedView` and `replaceText`.
+
 # Espresso-Helper 0.1.0 (2018-05-01)
 
 - Initial release.